Conditionally include go-tree-sitter in build process
